+++ NEWS.md
@@ -2,6 +2,7 @@
 
 New supported packages and models:
 
+* `survival::clogit`
 * `biglm`: The main quantities can be computed, but not the delta method standard errors. See https://github.com/vincentarelbundock/marginaleffects/issues/387
 
 New vignette: